fix(state-sync): Pass s3 credentials via a file
* config.json has a new field s3_credentials_file, which gives a relative location of a file containing access_key and secret_key
* Refactored state-viewer to use the same `ExternalConnection` abstraction as the rest of the code
