+++ chain/client/Cargo.toml
@@ -26,6 +26,7 @@ rand.workspace = true
 reed-solomon-erasure.workspace = true
 regex.workspace = true
 rust-s3.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 strum.workspace = true
 sysinfo.workspace = true
