+++ chain/client/src/sync/mod.rs
@@ -1,4 +1,5 @@
 pub mod block;
 pub mod epoch;
+pub mod external;
 pub mod header;
 pub mod state;
