+++ nearcore/src/lib.rs
@@ -348,6 +348,7 @@ pub fn start_with_config_and_synchronization(
             None
         };
 
+    let credentials_file = config.config.s3_credentials_file;
     let state_sync_dump_handle = spawn_state_sync_dump(
         &config.client_config,
         chain_genesis,
@@ -355,6 +356,7 @@ pub fn start_with_config_and_synchronization(
         shard_tracker,
         runtime,
         config.validator_signer.as_ref().map(|signer| signer.validator_id().clone()),
+        credentials_file.map(|filename| home_dir.join(filename)),
     )?;
 
     #[allow(unused_mut)]
