+++ nearcore/src/state_sync.rs
@@ -3,10 +3,12 @@ use borsh::BorshSerialize;
 use near_chain::types::RuntimeAdapter;
 use near_chain::{Chain, ChainGenesis, ChainStoreAccess, DoomslugThresholdMode, Error};
 use near_chain_configs::{ClientConfig, ExternalStorageLocation};
-use near_client::sync::state::{
-    external_storage_location, external_storage_location_directory, get_part_id_from_filename,
-    is_part_filename, ExternalConnection, StateSync, STATE_DUMP_ITERATION_TIME_LIMIT_SECS,
+use near_client::sync::external::{create_bucket_readwrite, external_storage_location};
+use near_client::sync::external::{
+    external_storage_location_directory, get_part_id_from_filename, is_part_filename,
+    ExternalConnection,
 };
+use near_client::sync::state::{StateSync, STATE_DUMP_ITERATION_TIME_LIMIT_SECS};
 use near_epoch_manager::shard_tracker::ShardTracker;
 use near_epoch_manager::EpochManagerAdapter;
 use near_primitives::hash::CryptoHash;
@@ -16,6 +18,7 @@ use near_primitives::types::{AccountId, EpochHeight, EpochId, ShardId, StateRoot
 use near_store::DBCol;
 use rand::{thread_rng, Rng};
 use std::collections::HashSet;
+use std::path::PathBuf;
 use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 use std::time::{Duration, Instant};
@@ -29,6 +32,7 @@ pub fn spawn_state_sync_dump(
     shard_tracker: ShardTracker,
     runtime: Arc<dyn RuntimeAdapter>,
     account_id: Option<AccountId>,
+    credentials_file: Option<PathBuf>,
 ) -> anyhow::Result<Option<StateSyncDumpHandle>> {
     let dump_config = if let Some(dump_config) = client_config.state_sync.dump.clone() {
         dump_config
@@ -40,23 +44,10 @@ pub fn spawn_state_sync_dump(
     tracing::info!(target: "state_sync_dump", "Spawning the state sync dump loop");
 
     let external = match dump_config.location {
-        ExternalStorageLocation::S3 { bucket, region } => {
-            // Credentials to establish a connection are taken from environment variables:
-            // * `AWS_ACCESS_KEY_ID`
-            // * `AWS_SECRET_ACCESS_KEY`
-            let creds = match s3::creds::Credentials::default() {
-                Ok(creds) => creds,
-                Err(err) => {
-                    tracing::error!(target: "state_sync_dump", "Failed to create a connection to S3. Did you provide environment variables AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY?");
-                    return Err(err.into());
-                }
-            };
-            let bucket = s3::Bucket::new(&bucket, region.parse::<s3::Region>()?, creds)?;
-            ExternalConnection::S3 { bucket: Arc::new(bucket) }
-        }
-        ExternalStorageLocation::Filesystem { root_dir } => {
-            ExternalConnection::Filesystem { root_dir }
-        }
+        ExternalStorageLocation::S3 { bucket, region } => ExternalConnection::S3{
+            bucket: Arc::new(create_bucket_readwrite(&bucket,&region,Duration::from_secs(30),credentials_file).expect("Failed to authenticate connection to S3. Did you forget to provide AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY in the environment."))
+        },
+        ExternalStorageLocation::Filesystem { root_dir } => ExternalConnection::Filesystem { root_dir },
     };
 
     // Determine how many threads to start.
