+++ tools/state-viewer/src/state_parts.rs
@@ -1,9 +1,11 @@
 use crate::epoch_info::iterate_and_filter;
 use borsh::BorshDeserialize;
 use near_chain::{Chain, ChainGenesis, ChainStoreAccess, DoomslugThresholdMode};
-use near_client::sync::state::{
-    get_num_parts_from_filename, is_part_filename, location_prefix, part_filename, StateSync,
+use near_client::sync::external::{
+    create_bucket_readonly, create_bucket_readwrite, external_storage_location,
+    external_storage_location_directory, get_num_parts_from_filename, ExternalConnection,
 };
+use near_client::sync::state::StateSync;
 use near_epoch_manager::shard_tracker::{ShardTracker, TrackedConfig};
 use near_epoch_manager::EpochManager;
 use near_primitives::challenge::PartialState;
@@ -16,12 +18,11 @@ use near_primitives_core::hash::CryptoHash;
 use near_primitives_core::types::{BlockHeight, EpochHeight, ShardId};
 use near_store::{PartialStorage, Store, Trie};
 use nearcore::{NearConfig, NightshadeRuntime};
-use s3::serde_types::ListBucketResult;
-use std::fs::DirEntry;
 use std::ops::Range;
 use std::path::{Path, PathBuf};
 use std::str::FromStr;
-use std::time::Instant;
+use std::sync::Arc;
+use std::time::{Duration, Instant};
 
 #[derive(clap::ValueEnum, Clone, Debug, Default)]
 pub(crate) enum LoadAction {
@@ -104,36 +105,86 @@ impl StatePartsSubCommand {
         )
         .unwrap();
         let chain_id = &near_config.genesis.config.chain_id;
-        match self {
-            StatePartsSubCommand::Load { action, state_root, part_id, epoch_selection } => {
-                load_state_parts(
-                    action,
-                    epoch_selection,
-                    shard_id,
-                    part_id,
-                    state_root,
-                    &mut chain,
-                    chain_id,
-                    store,
-                    Location::new(root_dir, (s3_bucket, s3_region)),
-                );
+        let sys = actix::System::new();
+        sys.block_on(async move {
+            let credentials_file =
+                near_config.config.s3_credentials_file.clone().map(|file| home_dir.join(file));
+            match self {
+                StatePartsSubCommand::Load { action, state_root, part_id, epoch_selection } => {
+                    let external = create_external_connection(
+                        root_dir,
+                        s3_bucket,
+                        s3_region,
+                        None,
+                        Mode::Readonly,
+                    );
+                    load_state_parts(
+                        action,
+                        epoch_selection,
+                        shard_id,
+                        part_id,
+                        state_root,
+                        &mut chain,
+                        chain_id,
+                        store,
+                        &external,
+                    )
+                    .await
+                }
+                StatePartsSubCommand::Dump { part_from, part_to, epoch_selection } => {
+                    let external = create_external_connection(
+                        root_dir,
+                        s3_bucket,
+                        s3_region,
+                        credentials_file,
+                        Mode::Readwrite,
+                    );
+                    dump_state_parts(
+                        epoch_selection,
+                        shard_id,
+                        part_from,
+                        part_to,
+                        &chain,
+                        chain_id,
+                        store,
+                        &external,
+                    )
+                    .await
+                }
+                StatePartsSubCommand::ReadStateHeader { epoch_selection } => {
+                    read_state_header(epoch_selection, shard_id, &chain, store)
+                }
             }
-            StatePartsSubCommand::Dump { part_from, part_to, epoch_selection } => {
-                dump_state_parts(
-                    epoch_selection,
-                    shard_id,
-                    part_from,
-                    part_to,
-                    &chain,
-                    chain_id,
-                    store,
-                    Location::new(root_dir, (s3_bucket, s3_region)),
-                );
-            }
-            StatePartsSubCommand::ReadStateHeader { epoch_selection } => {
-                read_state_header(epoch_selection, shard_id, &chain, store)
+        });
+        sys.run().unwrap();
+    }
+}
+
+enum Mode {
+    Readonly,
+    Readwrite,
+}
+
+fn create_external_connection(
+    root_dir: Option<PathBuf>,
+    bucket: Option<String>,
+    region: Option<String>,
+    credentials_file: Option<PathBuf>,
+    mode: Mode,
+) -> ExternalConnection {
+    if let Some(root_dir) = root_dir {
+        ExternalConnection::Filesystem { root_dir }
+    } else if let (Some(bucket), Some(region)) = (bucket, region) {
+        let bucket = match mode {
+            Mode::Readonly => create_bucket_readonly(&bucket, &region, Duration::from_secs(5)),
+            Mode::Readwrite => {
+                create_bucket_readwrite(&bucket, &region, Duration::from_secs(5), credentials_file)
             }
         }
+        .expect("Failed to create an S3 bucket");
+        ExternalConnection::S3 { bucket: Arc::new(bucket) }
+    } else {
+        panic!("Please provide --root-dir or both of --s3-bucket and --s3-region");
     }
 }
 
@@ -183,35 +234,6 @@ impl EpochSelection {
     }
 }
 
-enum Location {
-    Files(PathBuf),
-    S3 { bucket: String, region: String },
-}
-
-impl Location {
-    fn new(
-        root_dir: Option<PathBuf>,
-        s3_bucket_and_region: (Option<String>, Option<String>),
-    ) -> Self {
-        let (s3_bucket, s3_region) = s3_bucket_and_region;
-        assert_eq!(
-            s3_bucket.is_some(),
-            s3_region.is_some(),
-            "None or both of --s3-bucket and --s3-region need to be set"
-        );
-        assert_ne!(
-            root_dir.is_some(),
-            s3_bucket.is_some(),
-            "Only one of --root-dir and --s3 flags can be set"
-        );
-        if let Some(root_dir) = root_dir {
-            Location::Files(root_dir)
-        } else {
-            Location::S3 { bucket: s3_bucket.unwrap(), region: s3_region.unwrap() }
-        }
-    }
-}
-
 /// Returns block hash of some block of the given `epoch_info` epoch.
 fn get_any_block_hash_of_epoch(epoch_info: &EpochInfo, chain: &Chain) -> CryptoHash {
     let head = chain.store().head().unwrap();
@@ -243,7 +265,7 @@ fn get_any_block_hash_of_epoch(epoch_info: &EpochInfo, chain: &Chain) -> CryptoH
     }
 }
 
-fn load_state_parts(
+async fn load_state_parts(
     action: LoadAction,
     epoch_selection: EpochSelection,
     shard_id: ShardId,
@@ -252,7 +274,7 @@ fn load_state_parts(
     chain: &mut Chain,
     chain_id: &str,
     store: Store,
-    location: Location,
+    external: &ExternalConnection,
 ) {
     let epoch_id = epoch_selection.to_epoch_id(store, chain);
     let (state_root, epoch_height, epoch_id, sync_hash) =
@@ -272,10 +294,12 @@ fn load_state_parts(
             (state_root, epoch.epoch_height(), epoch_id, Some(sync_hash))
         };
 
-    let part_storage = get_state_part_reader(location, chain_id, &epoch_id, epoch_height, shard_id);
-
-    let num_parts = part_storage.num_parts();
-    assert_ne!(num_parts, 0, "Too few num_parts: {}", num_parts);
+    let directory_path =
+        external_storage_location_directory(chain_id, &epoch_id, epoch_height, shard_id);
+    let part_file_names = external.list_state_parts(shard_id, &directory_path).await.unwrap();
+    assert!(!part_file_names.is_empty());
+    let num_parts = part_file_names.len() as u64;
+    assert_eq!(Some(num_parts), get_num_parts_from_filename(&part_file_names[0]));
     let part_ids = get_part_ids(part_id, part_id.map(|x| x + 1), num_parts);
     tracing::info!(
         target: "state-parts",
@@ -291,7 +315,15 @@ fn load_state_parts(
     for part_id in part_ids {
         let timer = Instant::now();
         assert!(part_id < num_parts, "part_id: {}, num_parts: {}", part_id, num_parts);
-        let part = part_storage.read(part_id, num_parts);
+        let location = external_storage_location(
+            chain_id,
+            &epoch_id,
+            epoch_height,
+            shard_id,
+            part_id,
+            num_parts,
+        );
+        let part = external.get_part(shard_id, &location).await.unwrap();
 
         match action {
             LoadAction::Apply => {
@@ -337,7 +369,7 @@ fn print_state_part(state_root: &StateRoot, _part_id: PartId, data: &[u8]) {
     trie.print_recursive(&mut std::io::stdout().lock(), &state_root, u32::MAX);
 }
 
-fn dump_state_parts(
+async fn dump_state_parts(
     epoch_selection: EpochSelection,
     shard_id: ShardId,
     part_from: Option<u64>,
@@ -345,7 +377,7 @@ fn dump_state_parts(
     chain: &Chain,
     chain_id: &str,
     store: Store,
-    location: Location,
+    external: &ExternalConnection,
 ) {
     let epoch_id = epoch_selection.to_epoch_id(store, chain);
     let epoch = chain.epoch_manager.get_epoch_info(&epoch_id).unwrap();
@@ -372,9 +404,6 @@ fn dump_state_parts(
         "Dumping state as seen at the beginning of the specified epoch.",
     );
 
-    let part_storage =
-        get_state_part_writer(location, chain_id, &epoch_id, epoch.epoch_height(), shard_id);
-
     let timer = Instant::now();
     for part_id in part_ids {
         let timer = Instant::now();
@@ -388,7 +417,17 @@ fn dump_state_parts(
                 PartId::new(part_id, num_parts),
             )
             .unwrap();
-        part_storage.write(&state_part, part_id, num_parts);
+
+        let location = external_storage_location(
+            &chain_id,
+            &epoch_id,
+            epoch.epoch_height(),
+            shard_id,
+            part_id,
+            num_parts,
+        );
+        external.put_state_part(&state_part, shard_id, &location).await.unwrap();
+        // part_storage.write(&state_part, part_id, num_parts);
         let elapsed_sec = timer.elapsed().as_secs_f64();
         let first_state_record = get_first_state_record(&state_root, &state_part);
         tracing::info!(
@@ -434,209 +473,4 @@ fn read_state_header(
 
 fn get_part_ids(part_from: Option<u64>, part_to: Option<u64>, num_parts: u64) -> Range<u64> {
     part_from.unwrap_or(0)..part_to.unwrap_or(num_parts)
-}
-
-trait StatePartWriter {
-    fn write(&self, state_part: &[u8], part_id: u64, num_parts: u64);
-}
-
-trait StatePartReader {
-    fn read(&self, part_id: u64, num_parts: u64) -> Vec<u8>;
-    fn num_parts(&self) -> u64;
-}
-
-fn get_state_part_reader(
-    location: Location,
-    chain_id: &str,
-    epoch_id: &EpochId,
-    epoch_height: u64,
-    shard_id: ShardId,
-) -> Box<dyn StatePartReader> {
-    match location {
-        Location::Files(root_dir) => Box::new(FileSystemStorage::new(
-            root_dir,
-            false,
-            chain_id,
-            epoch_id,
-            epoch_height,
-            shard_id,
-        )),
-        Location::S3 { bucket, region } => {
-            Box::new(S3Storage::new(&bucket, &region, chain_id, epoch_id, epoch_height, shard_id))
-        }
-    }
-}
-
-fn get_state_part_writer(
-    location: Location,
-    chain_id: &str,
-    epoch_id: &EpochId,
-    epoch_height: u64,
-    shard_id: ShardId,
-) -> Box<dyn StatePartWriter> {
-    match location {
-        Location::Files(root_dir) => Box::new(FileSystemStorage::new(
-            root_dir,
-            true,
-            chain_id,
-            epoch_id,
-            epoch_height,
-            shard_id,
-        )),
-        Location::S3 { bucket, region } => {
-            Box::new(S3Storage::new(&bucket, &region, chain_id, epoch_id, epoch_height, shard_id))
-        }
-    }
-}
-
-struct FileSystemStorage {
-    state_parts_dir: PathBuf,
-}
-
-impl FileSystemStorage {
-    fn new(
-        root_dir: PathBuf,
-        create_dir: bool,
-        chain_id: &str,
-        epoch_id: &EpochId,
-        epoch_height: u64,
-        shard_id: u64,
-    ) -> Self {
-        let prefix = location_prefix(chain_id, epoch_height, epoch_id, shard_id);
-        let state_parts_dir = root_dir.join(&prefix);
-        if create_dir {
-            tracing::info!(target: "state-parts", ?root_dir, ?prefix, ?state_parts_dir, "Ensuring the directory exists");
-            std::fs::create_dir_all(&state_parts_dir).unwrap();
-        }
-        tracing::info!(target: "state-parts", ?state_parts_dir, "Initialized FileSystemStorage");
-        Self { state_parts_dir }
-    }
-
-    fn get_location(&self, part_id: u64, num_parts: u64) -> PathBuf {
-        self.state_parts_dir.join(part_filename(part_id, num_parts))
-    }
-}
-
-impl StatePartWriter for FileSystemStorage {
-    fn write(&self, state_part: &[u8], part_id: u64, num_parts: u64) {
-        let filename = self.get_location(part_id, num_parts);
-        std::fs::write(&filename, state_part).unwrap();
-        tracing::info!(target: "state-parts", part_id, part_length = state_part.len(), ?filename, "Wrote a state part to disk");
-    }
-}
-
-impl StatePartReader for FileSystemStorage {
-    fn read(&self, part_id: u64, num_parts: u64) -> Vec<u8> {
-        let filename = self.get_location(part_id, num_parts);
-        tracing::debug!(target: "state-parts", part_id, num_parts, ?filename, "Reading state part file");
-        std::fs::read(filename).unwrap()
-    }
-
-    fn num_parts(&self) -> u64 {
-        let paths = std::fs::read_dir(&self.state_parts_dir).unwrap();
-        let mut known_num_parts = None;
-        let num_files = paths
-            .filter(|path| {
-                let full_path = path.as_ref().unwrap();
-                tracing::debug!(target: "state-parts", ?full_path);
-                let filename = full_path.file_name().to_str().unwrap().to_string();
-                if let Some(num_parts) = get_num_parts_from_filename(&filename) {
-                    if let Some(known_num_parts) = known_num_parts {
-                        assert_eq!(known_num_parts, num_parts);
-                    }
-                    known_num_parts = Some(num_parts);
-                }
-                is_part_filename(&filename)
-            })
-            .collect::<Vec<std::io::Result<DirEntry>>>()
-            .len();
-        if known_num_parts != Some(num_files as u64) {
-            // This is expected when a user saves time and downloads a few parts instead of all parts.
-            tracing::warn!(target: "state-parts",
-                dir = ?self.state_parts_dir,
-                ?known_num_parts,
-                num_files,
-                "Filename indicates that number of files expected doesn't match the number of files available");
-        }
-        known_num_parts.unwrap()
-    }
-}
-
-struct S3Storage {
-    location: String,
-    bucket: s3::Bucket,
-}
-
-impl S3Storage {
-    fn new(
-        s3_bucket: &str,
-        s3_region: &str,
-        chain_id: &str,
-        epoch_id: &EpochId,
-        epoch_height: u64,
-        shard_id: u64,
-    ) -> Self {
-        let location = location_prefix(chain_id, epoch_height, epoch_id, shard_id);
-        let bucket = s3::Bucket::new(
-            s3_bucket,
-            s3_region.parse::<s3::Region>().unwrap(),
-            s3::creds::Credentials::default().unwrap(),
-        )
-        .unwrap();
-
-        tracing::info!(target: "state-parts", s3_bucket, s3_region, location, "Initialized an S3 bucket");
-        Self { location, bucket }
-    }
-
-    fn get_location(&self, part_id: u64, num_parts: u64) -> String {
-        format!("{}/{}", self.location, part_filename(part_id, num_parts))
-    }
-}
-
-impl StatePartWriter for S3Storage {
-    fn write(&self, state_part: &[u8], part_id: u64, num_parts: u64) {
-        let location = self.get_location(part_id, num_parts);
-        self.bucket.put_object_blocking(&location, state_part).unwrap();
-        tracing::info!(target: "state-parts", part_id, part_length = state_part.len(), ?location, "Wrote a state part to S3");
-    }
-}
-
-impl StatePartReader for S3Storage {
-    fn read(&self, part_id: u64, num_parts: u64) -> Vec<u8> {
-        let location = self.get_location(part_id, num_parts);
-        let response = self.bucket.get_object_blocking(location.clone()).unwrap();
-        tracing::info!(target: "state-parts", part_id, location, response_code = response.status_code(), "Got an object from S3");
-        assert_eq!(response.status_code(), 200);
-        response.into()
-    }
-
-    fn num_parts(&self) -> u64 {
-        let location = format!("{}/", self.location);
-        tracing::info!(target: "state-parts", location, "Reading an S3 bucket location to get the number of parts");
-        let list: Vec<ListBucketResult> =
-            self.bucket.list_blocking(location, Some("/".to_string())).unwrap();
-        assert_eq!(list.len(), 1);
-        let mut known_num_parts = None;
-        let num_objects = list[0]
-            .contents
-            .iter()
-            .filter(|object| {
-                let filename = Path::new(&object.key);
-                let filename = filename.file_name().unwrap().to_str().unwrap();
-                tracing::debug!(target: "state-parts", object_key = ?object.key, ?filename);
-                if let Some(num_parts) = get_num_parts_from_filename(filename) {
-                    if let Some(known_num_parts) = known_num_parts {
-                        assert_eq!(known_num_parts, num_parts);
-                    }
-                    known_num_parts = Some(num_parts);
-                }
-                is_part_filename(filename)
-            })
-            .collect::<Vec<&s3::serde_types::Object>>()
-            .len();
-        if let Some(known_num_parts) = known_num_parts {
-            assert_eq!(known_num_parts, num_objects as u64);
-        }
-        num_objects as u64
-    }
 }
